Extract RensaTracker from CoreField.
We'd like to define our own Tracker and use it. However, the current definition
of CoreField does not allow us to do so.

So, we extract Tracker from CoreField, and allow us to define our own Tracker.
The implementation of CoreField::simulate() is also extracted to
core_field_inl.h.