Introduce EnemyState.
We'd like to gaze enemy's rensa more correctly.
However, we cannot see 13th row of the enemy field, it's hard to determine
what the exact Rensa is. So, we'd like to hold all the possible Rensa in
EnemyState.

This patch just renames PlayerState to EnemyState. The upcoming patch will
introduce such gazer to see the possible Rensa.