Don't use static_cast to convert int to TLColor.
We'd like to use PuyoColor instead of TLColor in niina, too.
However, the primitive value of TLColor and PuyoColor is not the same.
To make it easy to understand the patch to convert TLColor to PuyoColor,
we'd like to remove static_cast beforehand.

Define NORMAL_TLCOLORS to convert int to TLColor, and use it.
Also, remove unnecessary TLColor to TLColor cast.